feat(collectiongallery): add an api that allows layouts to be customizable
